+++ CHANGELOG.md
@@ -13,6 +13,8 @@ If it saves you or your team time, please consider [sponsoring its development](
   and has been removed from the SDK.
 * Deprecated classes, methods and class constants have been removed.
 * Type declarations have been simplified to reduce runtime overhead (e.g., `Stringable|string` to `string`).
+* The transitional `Kreait\Firebase\Contract\Transitional\FederatedUserFetcher::getUserByProviderUid()` method
+  has been moved into the `Kreait\Firebase\Contract\Auth` interface
 
 See **[UPGRADE-8.0](UPGRADE-8.0.md) for more details on the changes between 7.x and 8.0.**